# Conflicts:
#	pharmacy_common/src/main/java/com/kgc/entity/OrderMedicine.java
#	pharmacy_inventory_view/src/components/CGSQaddOrder.vue

+++ pharmacy_common/src/main/java/com/kgc/entity/BaseMedicine.java
@@ -94,7 +94,7 @@ public class BaseMedicine implements Serializable {
     private double totalPrice;
 
     @TableField(exist=false)
-    private double providerId;
+    private Integer providerId;
 
     @TableField(exist=false)
     private double providerName;
